# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AppError:
    properties:
      codigo:
        type: integer
      mensagemErro:
        type: string
    type: object
...
  models.IntegrationItemError:
    properties:
      chave_item:
        type: string
      codigo_erro:
        type: string
      erro:
        type: string
    type: object
  models.IntegrationTransferDTO:
    properties:
      cnpj_destino:
        maxLength: 20
        type: string
      cnpj_origem:
        maxLength: 20
        type: string
      data_envio:
        type: string
      id:
        type: string
      id_loja_destino:
        type: string
      itens:
        items:
          $ref: '#/definitions/models.TransferLine'
        maxItems: 1000
        minItems: 1
        type: array
      numero:
        maxLength: 100
        type: string
      observacao_remetente:
        maxLength: 2000
        type: string
      usuario_remetente:
        maxLength: 100
        type: string
    required:
    - data_envio
    - id
    - itens
    - numero
    type: object
...
  models.Message:
    properties:
      acknowledgedAt:
        type: string
      attempts:
        type: integer
      claimToken:
        type: string
      createdAt:
        type: string
      destinationStoreId:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      lastAttemptAt:
        type: string
      lastError:
        type: string
      leaseUntil:
        type: string
      nextAttemptAt:
        type: string
      operation:
        type: string
      originStoreId:
        type: string
      payload:
        type: object
      status:
        type: string
      syncBatchId:
        type: string
      syncJobId:
        type: string
      transferId:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.Page:
    properties:
      items: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  models.Transfer:
    properties:
      createdAt:
        type: string
      destinationCnpj:
        type: string
      destinationStoreId:
        type: string
      id:
        type: string
      idEmpresa:
        type: string
      itemCount:
        type: integer
      numero:
        type: string
      originCnpj:
        type: string
      originStoreId:
        type: string
      receivedAt:
        type: string
      senderNote:
        type: string
      senderUsername:
        type: string
      sentAt:
        type: string
      status:
        type: string
      syncedAt:
        type: string
      updatedAt:
        type: string
    type: object
  models.TransferLine:
    properties:
      classe_terapeutica:
        type: string
      codigo_barras:
        type: string
      codigo_lote:
        type: string
      codigo_produto:
        type: string
      controlado:
        type: boolean
      data_fabricacao:
        type: string
      data_validade:
        type: string
      id:
        type: string
      id_produto:
        type: string
      lote:
        type: string
      lote_serial:
        type: string
      nome:
        maxLength: 255
        type: string
      preco_custo:
        type: number
      preco_venda:
        type: number
      quantidade:
        type: number
      registro_ms:
        type: string
      serial:
        type: string
      sku:
        type: string
      tabela_preco_informada:
        type: boolean
      tabela_preco_payload:
        type: string
      tipo_produto_sngpc:
        type: string
      unidade:
        type: string
      unidade_sngpc:
        type: string
      usa_lote:
        type: boolean
      usa_serial:
        type: boolean
      usa_tb_pc:
        type: string
      usa_tb_pc_informada:
        type: boolean
    required:
    - id
    - nome
    - quantidade
    type: object
  models.TransferStatusDTO:
    properties:
      erro:
        maxLength: 2000
        type: string
      erros_itens:
        items:
          $ref: '#/definitions/models.IntegrationItemError'
        type: array
      status:
        enum:
        - RECEIVED
        - FAILED
        - CANCELLED
        type: string
    required:
    - status
    type: object
  models.TransferSubmission:
    properties:
      duplicado:
        type: boolean
      id:
        type: string
      numero:
        type: string
      quantidade_itens:
        type: integer
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: StoreLink integrates stores. Administrators use JWT Bearer; each FNTS_Server
    installation uses its own fixed X-API-Key and database-validated X-Agent-ID.
  title: StoreLink API
  version: "1.0"
paths:
...
  /integracao/v1/transferencias:
    get:
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Rows per page, up to 100
        in: query
        name: pageSize
        type: integer
      - description: Transfer status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: List transfers addressed to this store
      tags:
      - Direct integration
    post:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Transfer and full line snapshots
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.IntegrationTransferDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransferSubmission'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TransferSubmission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Publish an idempotent transfer from FNTS
      tags:
      - Direct integration
  /integracao/v1/transferencias/{id}:
    get:
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Transfer UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Read a transfer involving this store
      tags:
      - Direct integration
    patch:
      consumes:
      - application/json
      parameters:
      - description: Store credential ID
        in: header
        name: X-Agent-ID
        required: true
        type: string
      - description: Transfer UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status and optional item errors
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TransferStatusDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transfer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - AgentAPIKey: []
      summary: Record transfer receipt, failure, or origin cancellation
      tags:
      - Direct integration
...
  /transfers:
    get:
      parameters:
      - description: Número da página
        in: query
        name: page
        type: integer
      - description: Itens por página
        in: query
        name: pageSize
        type: integer
      - description: Número da transferência
        in: query
        name: q
        type: string
      - description: Status da transferência
        in: query
        name: status
        type: string
      - description: UUID do grupo
        in: query
        name: groupId
        type: string
      - description: UUID da loja
        in: query
        name: storeId
        type: string
      - description: UUID da loja de origem
        in: query
        name: originStoreId
        type: string
      - description: UUID da loja de destino
        in: query
        name: destinationStoreId
        type: string
      - description: Data inicial
        in: query
        name: from
        type: string
      - description: Data final
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Page'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Lista transferências
      tags:
      - Transferências
  /transfers/{id}:
    get:
      description: Operação isolada por empresa. Identidade e autorização validadas
        pelo servidor.
      parameters:
      - description: UUID do recurso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Consulta transferência, itens e timeline
      tags:
      - Transferências
  /transfers/{id}/retry:
    post:
      description: Cria uma nova tentativa preservando transferência e itens originais.
      parameters:
      - description: UUID da transferência
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.AppError'
      security:
      - BearerAuth: []
      summary: Reenvia transferência após corrigir o catálogo de destino
      tags:
      - Transferências
securityDefinitions:
  AgentAPIKey:
    description: Per-agent fixed key shown once at provisioning or rotation.
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Bearer token obtained from /auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"